# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  auth.VerifyOTPRequest:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
  company.Company:
    properties:
      address:
        type: string
      bin:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone:
        type: string
      reputation_score:
        type: number
      reviewer_note:
        type: string
      status:
        $ref: '#/definitions/company.CompanyStatus'
      updated_at:
        type: string
      user_id:
        type: string
      verified:
        type: boolean
      website:
        type: string
    type: object
  company.CompanyStatus:
    enum:
    - pending
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusVerified
    - StatusRejected
  company.CreateCompanyRequest:
    properties:
      address:
        type: string
      bin:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  company.UpdateCompanyRequest:
    properties:
      address:
        type: string
      bin:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  errors.Code:
    enum:
    - NOT_FOUND
    - UNAUTHORIZED
    - VALIDATION
    - CONFLICT
    - INTERNAL
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeUnauthorized
    - CodeValidation
    - CodeConflict
    - CodeInternal
  errors.Error:
    properties:
      code:
        $ref: '#/definitions/errors.Code'
      details:
        additionalProperties: true
        type: object
      err: {}
      message:
        type: string
    type: object
  identity.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  identity.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  identity.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  identity.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  identity.User:
    properties:
      avatar_url:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
  identity.VerifyOTPRequest:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
  integrity.Company:
    properties:
      address:
        type: string
      bin:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone:
        type: string
      reputation_score:
        type: number
      reviewer_note:
        type: string
      status:
        $ref: '#/definitions/integrity.CompanyStatus'
      updated_at:
        type: string
      user_id:
        type: string
      verified:
        type: boolean
      website:
        type: string
    type: object
  integrity.CompanyStatus:
    enum:
    - pending
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusVerified
    - StatusRejected
  integrity.CreateCompanyRequest:
    properties:
      address:
        type: string
      bin:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  integrity.UpdateCompanyRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  jwt.TokenPair:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  marketplace.CreateReviewRequest:
    properties:
      comment:
        type: string
      rating:
        type: integer
      target_entity_id:
        type: string
      transaction_id:
        type: string
    type: object
  marketplace.ReputationScore:
    properties:
      average_rating:
        type: number
      entity_id:
        type: string
      review_count:
        type: integer
      tier:
        description: gold, silver, bronze, none
        type: string
    type: object
  marketplace.Review:
    properties:
      author_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: integer
      target_entity_id:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  profile.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  profile.User:
    properties:
      avatar_url:
        type: string
      company_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
  review.CreateReviewRequest:
    properties:
      comment:
        type: string
      rating:
        type: integer
      target_entity_id:
        type: string
      transaction_id:
        type: string
    type: object
  review.ReputationScore:
    properties:
      average_rating:
        type: number
      entity_id:
        type: string
      review_count:
        type: integer
      tier:
        description: gold, silver, bronze, none
        type: string
    type: object
  review.Review:
    properties:
      author_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: integer
      target_entity_id:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Industrix Industrial Equipment Marketplace API
  title: Industrix API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jwt.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Login user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Issue new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jwt.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with pending verification
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Register a new user
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Validates OTP sent to phone and issues initial tokens
      parameters:
      - description: OTP details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jwt.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Verify phone OTP
      tags:
      - auth
  /companies:
    post:
      consumes:
      - application/json
      description: Register a new company (requires verification flow to follow)
      parameters:
      - description: Company details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/company.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/company.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - companies
  /companies/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve company information by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.Company'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Get company details
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Update mutable fields of a company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/company.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update company details
      tags:
      - companies
  /reviews:
    post:
      consumes:
      - application/json
      description: Post a review and rating for a company or equipment
      parameters:
      - description: Review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/review.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/review.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - reviews
  /reviews/{entityID}:
    get:
      consumes:
      - application/json
      description: Get paginated reviews for an entity
      parameters:
      - description: Entity ID
        in: path
        name: entityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      summary: List reviews
      tags:
      - reviews
  /reviews/{entityID}/reputation:
    get:
      consumes:
      - application/json
      description: Get aggregated reputation metrics for an entity
      parameters:
      - description: Entity ID
        in: path
        name: entityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.ReputationScore'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Get reputation score
      tags:
      - reviews
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieve profile information for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update name and avatar of the authenticated user
      parameters:
      - description: Updated profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/profile.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"